requirements.txt
numpy
pandas
scikit-learn
streamlit
Pillow
